HLE/JIT: Invalidate cache lines by physical addresses
Virtual addressing can move around (MSR.IR can be turned on/off, BAT
and page tables can change) which makes it impossible to unhook
functions if the mappings change. Physical addresses are more
reliable since that's what the JIT Block Cache uses which is what we
need to invalidate.

The important effect of this change is that is severs the dependency
on the value of the MSR register. Functions can now be unhooked even
if the MSR.IR value differs from the value it had when the
registration was made (was true, now false; was false, now true).